Move api creation to a new repository
BaseActivity was only being used to instantiate the api object.
This commit removes the BaseActivity and create a MovieRepository
to handle the api.

Also, all data-related classes were moved to a 'repository' package.
