+++ app/src/main/java/com/arctouch/codechallenge/repository/MovieRepository.java
@@ -1,21 +1,21 @@
-package com.arctouch.codechallenge.base;
+package com.arctouch.codechallenge.repository;
 
-import android.support.v7.app.AppCompatActivity;
-
-import com.arctouch.codechallenge.api.TmdbApi;
+import com.arctouch.codechallenge.repository.api.TmdbApi;
 
 import okhttp3.OkHttpClient;
 import retrofit2.Retrofit;
 import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory;
 import retrofit2.converter.moshi.MoshiConverterFactory;
 
-public class BaseActivity extends AppCompatActivity {
+public class MovieRepository {
 
-    protected TmdbApi api = new Retrofit.Builder()
+      public TmdbApi api = new Retrofit.Builder()
             .baseUrl(TmdbApi.URL)
             .client(new OkHttpClient.Builder().build())
             .addConverterFactory(MoshiConverterFactory.create())
             .addCallAdapterFactory(RxJava2CallAdapterFactory.create())
             .build()
             .create(TmdbApi.class);
+
+
 }
