+++ app/src/main/java/com/arctouch/codechallenge/repository/model/GenreResponse.java
@@ -1,4 +1,4 @@
-package com.arctouch.codechallenge.model;
+package com.arctouch.codechallenge.repository.model;
 
 import java.util.List;
