+++ app/src/main/java/com/arctouch/codechallenge/splash/SplashActivity.java
@@ -3,24 +3,27 @@ package com.arctouch.codechallenge.splash;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
+import android.support.v7.app.AppCompatActivity;
 
 import com.arctouch.codechallenge.R;
-import com.arctouch.codechallenge.api.TmdbApi;
-import com.arctouch.codechallenge.base.BaseActivity;
-import com.arctouch.codechallenge.data.Cache;
+import com.arctouch.codechallenge.repository.MovieRepository;
+import com.arctouch.codechallenge.repository.api.TmdbApi;
+import com.arctouch.codechallenge.repository.data.Cache;
 import com.arctouch.codechallenge.home.HomeActivity;
 
 import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.schedulers.Schedulers;
 
-public class SplashActivity extends BaseActivity {
+public class SplashActivity extends AppCompatActivity {
+
+    private MovieRepository movieRepository = new MovieRepository();
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.splash_activity);
 
-        api.genres(TmdbApi.API_KEY, TmdbApi.DEFAULT_LANGUAGE)
+        movieRepository.api.genres(TmdbApi.API_KEY, TmdbApi.DEFAULT_LANGUAGE)
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
                 .subscribe(response -> {
